+++ rapidfuzz/distance/LCSseq.hpp
@@ -4,6 +4,7 @@
 #pragma once
 #include <rapidfuzz/distance/LCSseq_impl.hpp>
 
+#include <algorithm>
 #include <cmath>
 #include <limits>
 
@@ -91,7 +92,7 @@ private:
     template <typename InputIt2>
     int64_t maximum(detail::Range<InputIt2> s2) const
     {
-        return std::max(static_cast<int64_t>(s1.size()), s2.size());
+        return std::max(static_cast<ptrdiff_t>(s1.size()), s2.size());
     }
 
     template <typename InputIt2>
